refactor(emit): split class recovery erasure helpers
